Fix printing

+++ atl2.cpp
@@ -43,7 +43,7 @@ int main(int argc, char* argv[]) {
     c*=10;
   }
   for(int i=1;i<10;++i) {
-    breaks.insert((n*i)/10);
+    breaks.insert((static_cast<unsigned long long>(n)*i)/10);
   }
   breaks.insert(n);
